Move `DrumSampleTriggerSource` into its own class to avoid nested references
