+++ osu.Game.Rulesets.Taiko/UI/InputDrum.cs
@@ -8,7 +8,6 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Input.Bindings;
-using osu.Game.Audio;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Taiko.Objects;
 using osu.Game.Rulesets.UI;
@@ -201,24 +200,5 @@ namespace osu.Game.Rulesets.Taiko.UI
             }
         }
 
-        public class DrumSampleTriggerSource : GameplaySampleTriggerSource
-        {
-            public DrumSampleTriggerSource(HitObjectContainer hitObjectContainer)
-                : base(hitObjectContainer)
-            {
-            }
-
-            public void Play(HitType hitType)
-            {
-                var hitObject = GetMostValidObject();
-
-                if (hitObject == null)
-                    return;
-
-                PlaySamples(new ISampleInfo[] { hitObject.SampleControlPoint.GetSampleInfo(hitType == HitType.Rim ? HitSampleInfo.HIT_CLAP : HitSampleInfo.HIT_NORMAL) });
-            }
-
-            public override void Play() => throw new InvalidOperationException(@"Use override with HitType parameter instead");
-        }
     }
 }
